requirements.txt
asgiref==3.7.2
autopep8==2.0.0
azure-common==1.1.28
azure-core==1.28.0
azure-identity==1.12.0
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.2
azure-mgmt-storage==21.0.0
azure-storage-blob==12.17.0
azure-storage-common==2.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
Django==4.2.6
django-cors-headers==4.0.0
django-filter==23.2
django-storages==1.13.2
django-storages-azure==1.6.8
djangorestframework==3.14.0
dnspython==2.3.0
drf-yasg==1.21.7
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
portalocker==2.6.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.3
zipp==3.15.0